Moved more tests.
