+++ src/array/binary/from.rs
@@ -99,7 +99,6 @@ mod tests {
         let array = BinaryArray::<i32>::from(&[Some(b"hello".as_ref()), Some(b" ".as_ref()), None]);
 
         let a = array.validity().as_ref().unwrap();
-        assert_eq!(a.len(), 3);
-        assert_eq!(a.as_slice()[0], 0b00000011);
+        assert_eq!(a, &Bitmap::from([true, true, false]));
     }
 }
