+++ src/array/utf8/from.rs
@@ -265,8 +265,7 @@ mod tests {
         let array = Utf8Array::<i32>::from(&[Some("hello"), Some(" "), None]);
 
         let a = array.validity().as_ref().unwrap();
-        assert_eq!(a.len(), 3);
-        assert_eq!(a.as_slice()[0], 0b00000011);
+        assert_eq!(a, &Bitmap::from([true, true, false]));
     }
 
     #[test]
