+++ src/compute/aggregate/min_max.rs
@@ -108,9 +108,9 @@ where
     T: NativeType + Simd,
     T::Simd: SimdOrd<T>,
 {
-    if bitmap.offset() == 0 {
-        let validity_masks =
-            BitChunksExact::<<T::Simd as NativeSimd>::Chunk>::new(bitmap.as_slice(), bitmap.len());
+    let (slice, offset, length) = bitmap.as_slice();
+    if offset == 0 {
+        let validity_masks = BitChunksExact::<<T::Simd as NativeSimd>::Chunk>::new(slice, length);
         null_min_primitive_impl(values, validity_masks)
     } else {
         let validity_masks = bitmap.chunks::<<T::Simd as NativeSimd>::Chunk>();
@@ -125,9 +125,9 @@ where
     T: NativeType + Simd,
     T::Simd: SimdOrd<T>,
 {
-    if bitmap.offset() == 0 {
-        let validity_masks =
-            BitChunksExact::<<T::Simd as NativeSimd>::Chunk>::new(bitmap.as_slice(), bitmap.len());
+    let (slice, offset, length) = bitmap.as_slice();
+    if offset == 0 {
+        let validity_masks = BitChunksExact::<<T::Simd as NativeSimd>::Chunk>::new(slice, length);
         null_max_primitive_impl(values, validity_masks)
     } else {
         let validity_masks = bitmap.chunks::<<T::Simd as NativeSimd>::Chunk>();
