+++ src/compute/arithmetics/basic/sub.rs
@@ -414,19 +414,19 @@ mod tests {
         let (result, overflow) = overflowing_sub(&a, &b).unwrap();
         let expected = Int32Array::from(&[None, None, None, Some(0)]);
         assert_eq!(result, expected);
-        assert_eq!(overflow.as_slice()[0], 0b0000);
+        assert_eq!(overflow, Bitmap::from([false, false, false, false]));
 
         let a = Int8Array::from(&[Some(1i8), Some(-100i8)]);
         let b = Int8Array::from(&[Some(1i8), Some(100i8)]);
         let (result, overflow) = overflowing_sub(&a, &b).unwrap();
         let expected = Int8Array::from(&[Some(0i8), Some(56i8)]);
         assert_eq!(result, expected);
-        assert_eq!(overflow.as_slice()[0], 0b10);
+        assert_eq!(overflow, Bitmap::from([false, true]));
 
         // Trait testing
         let (result, overflow) = a.overflowing_sub(&b).unwrap();
         assert_eq!(result, expected);
-        assert_eq!(overflow.as_slice()[0], 0b10);
+        assert_eq!(overflow, Bitmap::from([false, true]));
     }
 
     #[test]
@@ -481,17 +481,17 @@ mod tests {
         let (result, overflow) = overflowing_sub_scalar(&a, &1i32);
         let expected = Int32Array::from(&[None, Some(5), None, Some(5)]);
         assert_eq!(result, expected);
-        assert_eq!(overflow.as_slice()[0], 0b0000);
+        assert_eq!(overflow, Bitmap::from([false, false, false, false]));
 
         let a = Int8Array::from(&[Some(1i8), Some(-100i8)]);
         let (result, overflow) = overflowing_sub_scalar(&a, &100i8);
         let expected = Int8Array::from(&[Some(-99i8), Some(56i8)]);
         assert_eq!(result, expected);
-        assert_eq!(overflow.as_slice()[0], 0b10);
+        assert_eq!(overflow, Bitmap::from([false, true]));
 
         // Trait testing
         let (result, overflow) = a.overflowing_sub(&100i8).unwrap();
         assert_eq!(result, expected);
-        assert_eq!(overflow.as_slice()[0], 0b10);
+        assert_eq!(overflow, Bitmap::from([false, true]));
     }
 }
