+++ tests/it/bitmap/mod.rs
@@ -1,0 +1,2 @@
+mod immutable;
+mod mutable;
