+++ tests/it/main.rs
@@ -1,1 +1,3 @@
+mod alloc;
+mod bitmap;
 mod buffer;
